feat: Add functions to manipulate calculator operations
Adds two functions to the calculator class to manipulate operations:

getLastOperation(): Returns the last value added to the operation array.
isOperator(): Verifies if a given value is a mathematical operator.
Also updates the addOperation() function to handle different types of input values, either adding them to the current operation or updating the last operator if it exists.